requirements.txt
validator-collection
pyserial


